+++ src/components/join/JoinForm.jsx
@@ -1,22 +1,13 @@
 import React, { useState, useEffect, useRef } from "react";
 import { useRecoilValue, useResetRecoilState } from "recoil";
-import { JoinButton, JoinTextField, Snackbar } from "../../components";
-import { FormField } from "../../styled";
+import { JoinButton, JoinFormField, Snackbar } from "../../components";
 import {
   joinInputs as joinInputsAtom,
   joinInputsSelector,
   joinValidationMessages,
   joinValidationMessagesSelector,
 } from "../../store";
-import {
-  emailFieldProps,
-  passwordFieldProps,
-  nicknameFieldProps,
-  passwordCheckFieldProps,
-} from "../../constants";
 import { hasNotEmptyStingValue, hasOnlyEmptyStingValue } from "../../util";
-
-const authType = "join";
 
 function JoinForm() {
   const joinInputs = useRecoilValue(joinInputsSelector);
@@ -43,8 +34,6 @@ function JoinForm() {
     setIsRightInput(isRightInput);
   }, [validationMessages, joinInputs]);
 
-  const isValid = (fieldName) => validationMessages[fieldName] === "";
-
   const failToJoin = (message) => {
     snackbarMessage.current = message;
     setIsOpenSnackbar(true);
@@ -52,24 +41,7 @@ function JoinForm() {
 
   return (
     <form>
-      <FormField>
-        <JoinTextField
-          fieldProps={nicknameFieldProps}
-          isValid={isValid("nickname")}
-        />
-        <JoinTextField
-          fieldProps={emailFieldProps(authType)}
-          isValid={isValid("email")}
-        />
-        <JoinTextField
-          fieldProps={passwordFieldProps(authType)}
-          isValid={isValid("password")}
-        />
-        <JoinTextField
-          fieldProps={passwordCheckFieldProps}
-          isValid={isValid("passwordCheck")}
-        />
-      </FormField>
+      <JoinFormField validationMessages={validationMessages} />
       <Snackbar
         open={isOpenSnackbar}
         setOpen={setIsOpenSnackbar}
