+++ src/components/join/index.js
@@ -1,4 +1,5 @@
 export { default as CancelJoinPopup } from "./CancelJoinPopup";
 export { default as JoinButton } from "./JoinButton";
 export { default as JoinForm } from "./JoinForm";
+export { default as JoinFormField } from "./JoinFormField";
 export { default as JoinTextField } from "./JoinTextField";
